[DROOLS-6445] Removed old KiePMMLPredicate factories
