+++ kie-pmml-trusty/kie-pmml-compiler/kie-pmml-compiler-commons/src/main/java/org/kie/pmml/compiler/commons/codegenfactories/KiePMMLPredicateFactory.java
@@ -63,5 +63,4 @@ public class KiePMMLPredicateFactory {
         }
     }
 
-
 }
